+++ cmd/kube-controller-manager/controller-manager.go
@@ -26,7 +26,7 @@ import (
 	"os"
 	"time"
 
-	"k8s.io/apiserver/pkg/util/logs"
+	"k8s.io/component-base/logs"
 	"k8s.io/kubernetes/cmd/kube-controller-manager/app"
 	_ "k8s.io/kubernetes/pkg/util/prometheusclientgo" // load all the prometheus client-go plugin
 	_ "k8s.io/kubernetes/pkg/version/prometheus"      // for version metric registration
