+++ staging/src/k8s.io/apiserver/pkg/util/globalflag/globalflags.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/spf13/pflag"
 
-	"k8s.io/apiserver/pkg/util/logs"
+	"k8s.io/component-base/logs"
 )
 
 // AddGlobalFlags explicitly registers flags that libraries (klog, verflag, etc.) register
