+++ staging/src/k8s.io/kube-aggregator/main.go
@@ -23,7 +23,7 @@ import (
 	"k8s.io/klog"
 
 	genericapiserver "k8s.io/apiserver/pkg/server"
-	"k8s.io/apiserver/pkg/util/logs"
+	"k8s.io/component-base/logs"
 	"k8s.io/kube-aggregator/pkg/cmd/server"
 
 	// force compilation of packages we'll later rely upon
